Final
I dont remember what I did casue I stopped committing accidentally but it looks like

I made a separate .h and .cpp file for the main arduino to call the servo functions for the buttons in labview just so that it did not clutter the main code.

I created all the code to match with the Version 4_23 File

I added the motor test file in there so we could directly work the motors with FSR to test them.

I added a convert Analog FSR function that would convert the FSR to return a force value but it was not used